fix(ci): restore the network-surface budgets #780's squash reverted
`main` has not compiled since #780. That PR branched before #779 landed
and its squash-merge reverted #779's work in every region where the two
branches' hunks overlapped — deleting definitions while leaving #779's
call sites in place. Seven E0425/E0609 errors in `stella-observatory`,
five more in `stella-serve`, which is also what took `docker-serve` down.

Restored, from the pre-squash blobs:

- `fsview.rs`: `MAX_DIR_ENTRIES`, `MAX_SNIPPET_READ_BYTES`, `read_head`
  and `file_len`. #780 kept its own doc rework of `user_config_dir`,
  which is what collided with them; that rework is preserved.
- `http.rs`: the `MAX_HEAD_BYTES` / `MAX_BODY_BYTES` / `READ_TIMEOUT`
  split. The revert had restored the older single `MAX_REQUEST_BYTES`,
  which no caller referenced.
- `server.rs`: the `counter` field, the `MAX_LIVE_TURNS` admission cap
  with its 429 + Retry-After, and unguessable turn ids (`new_turn_id`).

The two branches each added a *different* turn cap, so both are kept:
admission (refuse past 32 live) and eviction (drop settled past 64).
They compose, but only if admission counts live turns alone — counting
the whole registry would let a 64-entry settled backlog refuse every new
turn. `register_turn` therefore takes the spec and starts the session
under the same lock hold that checks the count, which is what makes the
cap hold against racing callers; `insert_started` is split out so the
eviction tests can still register already-finished sessions through the
real insert-and-evict path.

Two tests cover the restored cap, both mutation-checked: removing the
cap fails `a_turn_past_the_live_cap_is_refused`, and counting
`turns.len()` instead of live turns fails
`finished_turns_do_not_count_against_the_live_cap`.

An audit of the other six files #779 touched found no further behavioral
reverts, but two documentation ones. `serve`'s `# Operational limits`
block still advertised "no admission control", "turn ids are sequential"
and "no read timeout" — telling operators to compensate for three
measures that exist a few hundred lines below it. `stella-serve/README.md`
carried #780's re-inserted 1 MiB bullet directly above #779's 64 KiB/8 MiB
replacement, so the file documented both. Both now match the code.

Separately, `stella-observatory/src/lib.rs` had crossed the 1500-line
ratchet at 1524. Its 1054-line test module moves to `tests.rs`, matching
how `deck_render`/`deck_ui` already split theirs; lib.rs is now 469
lines. The move is verbatim — rustdoc re-indents the code, and the
multi-line SQL and JSONL literals are byte-identical.

Every gate green locally: fmt, clippy, doc, the three script guards, the
release smoke build, and 4045 tests across 63 binaries.
